refactor(): some heavy duty restructuring * also added experimental saving of some settings
