+++ packages/jscad-desktop/index.html
@@ -60,7 +60,6 @@
             <option value='light'>Light Theme</option>
           </select>
           <span id='exports'>
-            
             <select id='exportFormats'>
             </select>
             <input type='button' value="export" id="exportBtn"/>
