+++ packages/jscad-desktop/package.json
@@ -20,7 +20,8 @@
     "@jscad/stl-serializer": "^0.1.1",
     "@jscad/svg-serializer": "^0.1.1",
     "@jscad/x3d-serializer": "^0.1.1",
-    "blob-to-buffer": "^1.2.6"
+    "blob-to-buffer": "^1.2.6",
+    "electron-store": "^1.3.0"
   },
   "devDependencies": {
     "browserify": "^14.5.0",
