+++ packages/jscad-desktop/src/core/getParameterDefinitionsCLI.js
