+++ packages/jscad-desktop/src/core/rebuildSolid.js
@@ -1,0 +1,12 @@
+
+const {toArray} = require('../utils')
+const getParamValues = require('./getParamValues')
+
+const rebuildSolid = (jscadScript, paramControls) => {
+  console.log('rebuilding')
+  let params = getParamValues(paramControls)
+  const csgs = toArray(jscadScript(params))
+  return csgs
+}
+
+module.exports = {rebuildSolid}
